[MINOR]add styles for todo app
