+++ samples/subapp2-todo-app/package.json
@@ -38,6 +38,7 @@
   "devDependencies": {
     "@types/node": "^14.14.6",
     "@xarc/app-dev": "^8.2.0",
+    "classnames": "^2.2.6",
     "ts-node": "^9.0.0",
     "typescript": "^4.0.3"
   }
