+++ samples/subapp2-todo-app/src/todo/component/todo.tsx
@@ -1,12 +1,14 @@
 import { React, ReactSubApp, AppContext } from "@xarc/react";
 import { connect, reduxFeature } from "@xarc/react-redux";
 import { toggleTodo } from "../redux/action";
+import custom from "../styles/bootstrap.css";
 
 const Todo = ({ todo, dispatch }) => {
   console.log("todo");
   console.log(todo);
   return (
     <li
+      className={custom["list-group-item"]}
       onClick={() => {
         console.log(todo.id);
         dispatch(toggleTodo(todo.id));
