+++ samples/subapp2-todo-app/src/todo/component/todoapp.tsx
@@ -1,4 +1,6 @@
 import { React, ReactSubApp, createDynamicComponent } from "@xarc/react";
+import classNames from "classnames";
+import custom from "../styles/bootstrap.css";
 
 export const TodoInput = createDynamicComponent(
   {
@@ -17,7 +19,7 @@ export const TodoList = createDynamicComponent(
 function TodoApp() {
   return (
     <div>
-      <h1>My To-do List</h1>
+      <h1 className={classNames(custom["page-header"], custom["text-center"])}>My To-do List</h1>
       <TodoInput />
       <TodoList />
     </div>
