--- samples/subapp2-todo-app/src/todo/redux/store.js
@@ -1,4 +1,0 @@
-import { createStore } from "redux";
-import { todoReducer } from "./reducer";
-
-export default createStore(todoReducer);
